+++ HISTORY.rst
@@ -12,6 +12,7 @@ HEAD (unreleased)
 - Fixing email sending (obtaining of address).
 - More human-oriented rendering and editing of bases mask strings.
 - Update cache when registering new histogram stats (#25).
+- Fixing search for messages (#15).
 
 ------
 v0.1.1